Warn if any segment has a running basebackup (#13620)
gprecoverseg, gpmovemirrors should show warning and gpexpand should raise
an exception if there is already a running instance of pg_basebackup for
the segment that needs to be recovered/moved.

Fix:
Create a list segments_with_running_basebackup of segment dbids for which
pg_basebackup is running and while creating the recovery triplets, check
if the segment to be recovered/moved/added(expansion segments) is already
present in the segments_with_running_basebackup list. If that's the case
give warning or raise exception and terminate the utility execution.

* Behaviour of gprecoverseg -
  Warning message is logged containing list of segment contentIds with
  running basebackup and for the rest of the segments gprecoverseg will
  proceed with recovery

* Behaviour of gpmovemirrors -
  Warning message is logged containing list of segment contentIds with
  running basebackup and for the rest of the mirrors gpmovemirrors will
  proceed with movement

* Behaviour of gpexpand -
  Exception is raised and execution is stopped immediately if segments to
  be added have running pg_basebackup

* Behaviour of gpaddmirrors -
  As gpaddmirrors does not use force-overwrite flag with pg_basebackup to
  sync newly added mirrors with primary, the check is not required here.
  Have added behave test case to ensure it fails when mirror directories
  are not empty

Limitation of above fix:
To get the list of segments with running pg_basebackup we are querying
gp_stat_replication table. At present this table shows only content ID
of the source segment for any running pg_basebackup process and no
information about the target segment/DataDirectory

Added behave testcase for gprecoverseg, gpmovemirrors

+++ gpMgmt/test/behave/mgmt_utils/environment.py
@@ -149,7 +149,7 @@ def after_scenario(context, scenario):
         return
 
     tags_to_cleanup = ['gpmovemirrors', 'gpssh-exkeys']
-    if set(context.feature.tags).intersection(tags_to_cleanup):
+    if set(context.feature.tags).intersection(tags_to_cleanup) and "skip_cleanup" not in scenario.effective_tags:
         if 'temp_base_dir' in context and os.path.exists(context.temp_base_dir):
             os.chmod(context.temp_base_dir, 0o700)
             shutil.rmtree(context.temp_base_dir)
